refactor(thread): reorganize code around the new thread mod
