+++ crates/matrix-sdk/src/event_cache/room/threads/serde_helpers.rs
@@ -1,0 +1,54 @@
+// Copyright 2025 The Matrix.org Foundation C.I.C.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use ruma::{events::relation::BundledThread, OwnedEventId};
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct Relations {
+    #[serde(rename = "m.thread")]
+    pub thread: Option<Box<BundledThread>>,
+}
+
+#[derive(Deserialize)]
+pub struct Unsigned {
+    #[serde(rename = "m.relations")]
+    pub relations: Option<Relations>,
+}
+
+#[derive(Deserialize)]
+enum RelationsType {
+    #[serde(rename = "m.thread")]
+    Thread,
+}
+
+#[derive(Deserialize)]
+struct RelatesTo {
+    #[serde(rename = "rel_type")]
+    rel_type: RelationsType,
+    #[serde(rename = "event_id")]
+    event_id: Option<OwnedEventId>,
+}
+
+#[derive(Deserialize)]
+pub struct SimplifiedContent {
+    #[serde(rename = "m.relates_to")]
+    relates_to: Option<RelatesTo>,
+}
+
+impl SimplifiedContent {
+    pub fn thread_root(self) -> Option<OwnedEventId> {
+        self.relates_to?.event_id
+    }
+}
